# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type MarketAssetStatus = typeof MarketAssetStatus[keyof typeof MarketAssetStatus];


export const MarketAssetStatus = {
  up: 'up',
  down: 'down',
  neutral: 'neutral',
} as const;

export interface MarketAsset {
  id: string;
  name: string;
  symbol: string;
  value: number;
  formattedValue: string;
  change: number;
  changePercent: number;
  lastUpdated: string;
  source: string;
  status: MarketAssetStatus;
}

export type MarketOverviewGeopoliticalRisk = typeof MarketOverviewGeopoliticalRisk[keyof typeof MarketOverviewGeopoliticalRisk];


export const MarketOverviewGeopoliticalRisk = {
  low: 'low',
  moderate: 'moderate',
  elevated: 'elevated',
  high: 'high',
} as const;

export interface MarketOverview {
  updatedAt: string;
  assets: MarketAsset[];
  geopoliticalRisk: MarketOverviewGeopoliticalRisk;
  /**
     * @minimum 0
     * @maximum 100
     */
  geopoliticalRiskScore: number;
  geopoliticalRiskLabel: string;
  riskDrivers: string[];
}

export type NewsImpactDirection = typeof NewsImpactDirection[keyof typeof NewsImpactDirection];


export const NewsImpactDirection = {
  'strong-up': 'strong-up',
  up: 'up',
  neutral: 'neutral',
  down: 'down',
  'strong-down': 'strong-down',
} as const;

export interface NewsImpact {
  asset: string;
  direction: NewsImpactDirection;
}

export interface NewsArticle {
  id: string;
  headline: string;
  source: string;
  url: string;
  publishedAt: string;
  /** @nullable */
  imageUrl: string | null;
  affectedAssets: string[];
  explanation: string;
  impacts: NewsImpact[];
}

export interface NewsResponse {
  articles: NewsArticle[];
  updatedAt: string;
  query: string;
}

export type MarketComparisonFavored = typeof MarketComparisonFavored[keyof typeof MarketComparisonFavored];


export const MarketComparisonFavored = {
  gold: 'gold',
  stocks: 'stocks',
  balanced: 'balanced',
} as const;

export type MarketComparisonRiskLevel = typeof MarketComparisonRiskLevel[keyof typeof MarketComparisonRiskLevel];


export const MarketComparisonRiskLevel = {
  low: 'low',
  moderate: 'moderate',
  elevated: 'elevated',
  high: 'high',
} as const;

export interface MarketComparison {
  updatedAt: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  goldScore: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  stocksScore: number;
  favored: MarketComparisonFavored;
  riskLevel: MarketComparisonRiskLevel;
  supportiveFactors: string[];
  negativeFactors: string[];
  explanation: string;
}

export type GetNewsParams = {
/**
 * @minimum 1
 * @maximum 20
 */
limit?: number;
};