# lib/api-zod/src/generated/types/marketAsset.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MarketAssetStatus } from './marketAssetStatus';

export interface MarketAsset {
  id: string;
  name: string;
  symbol: string;
  value: number;
  formattedValue: string;
  change: number;
  changePercent: number;
  lastUpdated: string;
  source: string;
  status: MarketAssetStatus;
}