# lib/api-zod/src/generated/types/marketComparison.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MarketComparisonFavored } from './marketComparisonFavored';
import type { MarketComparisonRiskLevel } from './marketComparisonRiskLevel';

export interface MarketComparison {
  updatedAt: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  goldScore: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  stocksScore: number;
  favored: MarketComparisonFavored;
  riskLevel: MarketComparisonRiskLevel;
  supportiveFactors: string[];
  negativeFactors: string[];
  explanation: string;
}