# lib/api-zod/src/generated/types/marketOverview.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MarketAsset } from './marketAsset';
import type { MarketOverviewGeopoliticalRisk } from './marketOverviewGeopoliticalRisk';

export interface MarketOverview {
  updatedAt: string;
  assets: MarketAsset[];
  geopoliticalRisk: MarketOverviewGeopoliticalRisk;
  /**
     * @minimum 0
     * @maximum 100
     */
  geopoliticalRiskScore: number;
  geopoliticalRiskLabel: string;
  riskDrivers: string[];
}