# lib/api-zod/src/generated/types/newsArticle.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NewsImpact } from './newsImpact';

export interface NewsArticle {
  id: string;
  headline: string;
  source: string;
  url: string;
  publishedAt: string;
  /** @nullable */
  imageUrl: string | null;
  affectedAssets: string[];
  explanation: string;
  impacts: NewsImpact[];
}