# lib/api-zod/src/generated/types/newsImpact.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NewsImpactDirection } from './newsImpactDirection';

export interface NewsImpact {
  asset: string;
  direction: NewsImpactDirection;
}