# lib/api-zod/src/generated/types/newsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NewsArticle } from './newsArticle';

export interface NewsResponse {
  articles: NewsArticle[];
  updatedAt: string;
  query: string;
}